refactor: extract 4 zero-shared-data panels from PanelViews.tsx
Batch 1 of the per-panel split for issue #144: adaptive_test,
test_history, worksheet_templates, system_settings — the 4 panels that
depend on neither usePanelData nor any local UI state in PanelViews.tsx,
just their own static/mock content.

- New panels/AdaptiveTestPanel.tsx, TestHistoryPanel.tsx,
  WorksheetTemplatesPanel.tsx, SystemSettingsPanel.tsx — each moved
  verbatim, along with its own mock data array (DIAGNOSTIC_HISTORY,
  WS_TEMPLATES, SYSTEM_LOGS_MOCK) that only that panel used.
- PanelViews.tsx's 4 if-blocks replaced with one-line router calls
  (if (panel === 'x') return <XPanel />;).
- CONTENT_ITEMS (the dead mock array flagged in the #144 design
  proposal) deliberately left untouched — out of scope for this PR.

Verified: tsc --noEmit clean, vite build succeeds, byte-diff of every
moved block, and a live click-through of all 4 panels against a
scratch-seeded backend (Superadmin for Worksheet Templates/System
Settings, Teacher for Adaptive Test/Test History) — all render
identically to before.
